d3_fastapi/pyproject.toml

[project]
name = ""
version = ""
description = ""
authors = [
    {name = "MarlonOu", email = "jj80868588@gmail.com"},
]
dependencies = [
    "mysql-connector-python",
    "uvicorn",
    "fastapi",
    "python-jose",
    "python-dotenv",
    "python-multipart",
    "pydantic==1.10.9",
    "sqlalchemy>=2.0.19",
    "sqlalchemy",
]
requires-python = ">=3.11"
license = {text = "MIT"}
readme = "README.md"



[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"
